+++ app/src/main/java/com/magic/haptic/util/TriggerCountFormatter.kt
@@ -1,0 +1,7 @@
+package com.magic.haptic.util
+
+object TriggerCountFormatter {
+    fun format(count: Int): String {
+        return "Triggers this session: $count"
+    }
+}